Move back to single file structure
