+++ pyproject.toml
@@ -17,11 +17,8 @@ classifiers = [
   "Development Status :: 3 - Alpha",
   "Environment :: Console",
   "Intended Audience :: Developers",
-  "License :: OSI Approved :: Apache Software License",
+  "License :: OSI Approved :: MIT License",
   "Natural Language :: English",
-  "Operating System :: MacOS :: MacOS X",
-  "Operating System :: POSIX :: Linux",
-  "Operating System :: Microsoft :: Windows",
   "Operating System :: OS Independent",
   "Programming Language :: Python :: 3",
   "Programming Language :: Python :: 3 :: Only",
@@ -41,14 +38,3 @@ dev = [
 
 [project.urls]
 source = "https://github.com/brettcannon/microvenv"
-
-[project.scripts]
-microvenv = "microvenv.main:main"
-
-[tool.pytest.ini_options]
-minversion = "7.0"
-addopts = [ "-ra", "--strict-markers", "--strict-config" ]
-xfail_strict = true
-log_cli_level = "info"
-pythonpath = [ "src" ]
-testpaths = [ "tests" ]
